+++ mlcolvar/core/transform/tools/normalization.py
@@ -200,7 +200,7 @@ def test_normalization():
     X = torch.randn((100, in_features)) * 10
 
     # get stats
-    from mlcolvar.core.transform.tools.utils import Statistics
+    from mlcolvar.core.transform.utils import Statistics
 
     stats = Statistics(X).to_dict()
     norm = Normalization(in_features, mean=stats["mean"], range=stats["std"])
